feat: Add Binary Tree Right Side View algorithm (#5)
* feat: Add Binary Tree Right Side View algorithm

Added the Binary Tree Right Side View algorithm to the BFS fundamentals section
of the Algo Monster repository. This algorithm is used to get the last element
of the level order traversal.

* feat: Add Binary Tree Right Side View algorithm

Added a new algorithm to get the last element of level order traversal in a
binary tree. The algorithm implements the Binary Tree Right Side View.